Analysis Batch GUI run) Refactoring each ExpDb Input Nodes - Common code for each expdb node
